requirements-minimal.txt
# Minimal requirements for Phase 1A Authentication Testing
fastapi==0.115.13
uvicorn[standard]==0.34.3
sqlalchemy==2.0.41
asyncpg==0.30.0
supabase==2.15.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.1.0
httpx==0.28.1
python-multipart==0.0.20